Fix fake BN core generation
Turns out that comments are valid things to put at the end of lines, so
I've just stopped looking at the end of lines!

Also, s/BN_CORE_LIBRARY/CORE_LIBRARY to account for the name used in
FindBinaryNinjaCore.cmake. Definitely not depending on implementation
details here...

Also, we weren't rebuilding fake libbinaryninjacore.dylib on update of
gen_lib.py, so we do that now.